Added the functions into modules
