+++ src/index.js
@@ -1,6 +1,5 @@
 import './style.css';
-
-let url = "https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/c18mtm405VaHSCO6kXTh/scores"
+import { submit, get } from './modules/getAndAdd';
 
 // let submit = (user, score) => fetch(url, {
 //   method: "POST",
@@ -17,22 +16,6 @@ let url = "https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/
 //     return res.json()})
 //   .then(data => console.log(data))
 
-const submit = async (user, score) => { 
-  let response = await fetch( url, {
-  method: "POST",
-  headers: {
-    "Content-Type": "application/json",
-  },
-  body: JSON.stringify({
-    user,score
-  })
-})
-
- 
-  let data = await response.json()
-  console.log(data)
-  return data
-}
 
 let submitForm = document.querySelector(".submit")
 
@@ -62,18 +45,5 @@ submitForm.addEventListener("submit", (e) => {
       
      
     // })
-    
-  const get = async () => {
-    const res = await fetch(url);
-
-    const data = await res.json();
-    data.result.forEach(user => {
-
-      const newname = `<li>${user.user}:${user.score}</li>`
-      const list = document.querySelector("ul")
-      list.innerHTML += newname
-
-    });
-  }
 
   get()
